done most part of dice_roll

+++ src/dicebag.hpp
@@ -16,7 +16,7 @@ public:
   DiceBag(size_type green_amount, size_type yellow_amount, size_type red_amount, std::string green_faces, std::string yellow_faces, std::string red_faces);
 
   //METHODS:
-  std::string sort_dices(int num_dice = 3);
+  std::string sort_dices(size_t num_dice = 3);
   bool lower_than_3_dices();
   void refill_bag();
